added signed ints

+++ functions_to_implement.cpp
@@ -36,7 +36,7 @@ int Sum(std::vector<int> nums)
 {
 	int sum = 0;
 
-	for(int i = 0; i < nums.size(); i++)
+	for(int i = 0; (int)i < (int)nums.size(); i++)
 	{
 		sum += nums[i];
 	}
@@ -49,7 +49,7 @@ int Product(std::vector<int> nums)
 {
 	int product = 1;
 
-	for(int i = 0; i < nums.size(); i++)
+	for(int i = 0; (int)i < (int)nums.size(); i++)
 	{
 		product *= nums[i];
 	}
